use the net beans for this project
The .form are the frontend GUI created to interact and .java files are the codes for every form named similarly. Some pages you need to add library packages based on errors. you can google the errors and find the solutions. the other .png files are simply added for icons in the buttons etc. you can neglect them and use your own icons. all the best